# docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: TekClinic REST API
  description: API of TekClinic
  version: 1.0.0
servers:
- url: /
paths:
...
  /doctors:
    get:
      tags:
      - Doctor
      description: returns all registered doctors
      operationId: getDoctors
      parameters:
      - name: skip
        in: query
        description: number of doctors to skip
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        description: max number of doctors to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 1
          type: integer
          format: int32
          default: 20
      - name: search
        in: query
        description: Search term for filtering. It searches name, phone number, specialities and special note. All words in the search term must appear in one of the fields as a prefix.
        required: false
        style: form
        explode: true
        schema:
          maxLength: 100
          minLength: 1
          type: string
      responses:
        "200":
          description: all the doctors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamedAPIResourceList'
              example:
                count: 12
                results:
                - id: 1
                  name: doctor
                  url: http://api.tekclinic.org/doctors/1/
                - id: 2
                  name: doctor
                  url: http://api.tekclinic.org/doctors/2/
    post:
      tags:
        - Doctor
      operationId: createDoctor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorBase'
      responses:
        "201":
          description: id of the created doctor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
  /doctors/{id}:
    get:
      tags:
      - Doctor
      description: returns a specific doctor
      operationId: getDoctor
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: doctor associated with the id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
    put:
      tags:
        - Doctors
      description: updates a specific doctor
      operationId: updateDoctor
      parameters:
        - name: id
          in: path
          description: id of the doctor to be updated
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorUpdate'
      responses:
        "200":
          description: id of the updated doctor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
    delete:
      tags:
        - Doctor
      description: deletes a doctor
      operationId: deleteDoctor
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: doctor deleted successfully
  /appointments:
    get:
      tags:
      - Appointment
      description: returns all registered appointments
      operationId: getAppointments
      parameters:
      - name: skip
        in: query
        description: number of appointments to skip
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        description: max number of appointments to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 1
          type: integer
          format: int32
          default: 20
      - name: date
        in: query
        description: return appointments only in the specified day
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
        example: 2024-11-23
      - name: doctor
        in: query
        description: return appointments only of the specified doctor (id)
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
        example: 11
      - name: patient
        in: query
        description: return appointments only of the specified patient (id)
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
        example: 21
      responses:
        "200":
          description: all the appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamedAPIResourceList'
              example:
                count: 12
                next: http://api.tekclinic.org/appointments/?limit=20&offset=20&date=2024-11-23
                results:
                - id: 23
                  name: appointment
                  url: http://api.tekclinic.org/appointments/23/
                - id: 122
                  name: appointment
                  url: http://api.tekclinic.org/appointments/122/
    post:
      tags:
      - Appointment
      description: "creates an empty appointment of the doctor corresponding to the *doctor_id*. If *patient_id* supplied, also assigns the patient to this appointment."
      operationId: createAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentBase'
      responses:
        "201":
          description: id of the created appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
  /appointments/{id}:
    get:
      tags:
      - Appointment
      description: returns a specific appointment
      operationId: getAppointment
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: appointment associated with the id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
    put:
      tags:
      - Appointment
      description: updates a specific appointment
      operationId: updateAppointment
      parameters:
      - name: id
        in: path
        description: id of the appointment to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        "200":
          description: id of the updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
    delete:
      tags:
      - Appointment
      description: deletes an appointment
      operationId: deleteAppointment
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: appointment delete successfully
  /appointments/{id}/patient:
    put:
      tags:
      - Appointment
      description: assigns a patient to the appointment
      operationId: assignPatientToAppointment
      parameters:
      - name: id
        in: path
        description: id of the appointment
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/PatientIdHolder'
      requestBody:
        description: patient that will be assigned to the appointment
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientIdHolder'
            example:
              patient_id: 12
      responses:
        "200":
          description: appointment associated with the id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientIdHolder'
    delete:
      tags:
      - Appointment
      description: unassign any patient assigned to the appointment
      operationId: clearPatientFromAppointment
      parameters:
      - name: id
        in: path
        description: id of the appointment
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: previously assigned patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientIdHolder'
components:
  schemas:
    NamedAPIResourceList:
      required:
      - count
      - next
      - previous
      - results
      type: object
      properties:
        count:
          type: number
          description: total number of resources available from this API.
          format: int32
        next:
          type: string
          description: URL of the next page in the list.
          format: uri
          nullable: true
        previous:
          type: string
          description: URL of the previous page in the list.
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/NamedAPIResource'
    NamedAPIResource:
      required:
      - id
      - name
      - url
      type: object
      properties:
        id:
          type: integer
          description: identifier of the referenced resource.
          format: int32
        name:
          type: string
          description: type name of the referenced resource.
        url:
          type: string
          description: URL of the referenced resource.
          format: uri
          deprecated: true
...
    DoctorBase:
      required:
      - name
      - phone_number
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: full name of doctor.
          example: Moshe Strauss
        gender:
          type: string
          description: gender of the doctor.
          example: male
          default: unspecified
          enum:
          - unspecified
          - male
          - female
        phone_number:
          type: string
          description: "phone number of doctor, that includes country code."
          example: +972501131591
        specialities:
          maxItems: 30
          type: array
          description: specialities of the doctor.
          example:
          - Family
          - OBGYN
          items:
            type: string
          default: []
        special_note:
          maxLength: 500
          type: string
          description: additional notes.
          example: Works only with males
    Doctor:
      allOf:
        - required:
            - active
            - gender
            - id
            - name
            - phone_number
            - specialities
          type: object
          properties:
            id:
              type: integer
              description: identifier of the doctor.
              format: int32
              example: 13
            active:
              type: boolean
              description: whether the doctor still works in the clinic.
              example: true
              default: true
        - $ref: '#/components/schemas/DoctorBase'
    DoctorUpdate:
      allOf:
        - $ref: '#/components/schemas/DoctorBase'
        - required:
            - active
            - name
            - gender
            - phone_number
            - specialities
            - special_note
          type: object
          properties:
            active:
              type: boolean
              description: whether the doctor still works in the clinic.
              example: true
              default: true
    AppointmentBase:
      required:
      - doctor_id
      - end_time
      - start_time
      type: object
      properties:
        patient_id:
          type: integer
          description: patient assigned to the appointment
          format: int32
          example: 123
        doctor_id:
          type: integer
          description: doctor assigned to the appointment
          format: int32
          example: 11
        start_time:
          type: string
          description: start time of the appointment
          format: date-time
          example: 2024-03-20T14:30:00Z
        end_time:
          type: string
          description: end time of the appointment
          format: date-time
          example: 2024-03-20T15:30:00Z
    Appointment:
      allOf:
      - $ref: '#/components/schemas/AppointmentBase'
      - required:
        - approved_by_patient
        - doctor_id
        - end_time
        - id
        - start_time
        - visited
        type: object
        properties:
          id:
            type: integer
            description: identifier of the appointment.
            format: int32
            example: 13
          approved_by_patient:
            type: boolean
            description: whether the assigned patient approved that he/she will arrive
            example: true
            default: false
          visited:
            type: boolean
            description: whether the assigned patient come to the appointment
            default: false
    AppointmentUpdate:
      allOf:
      - $ref: '#/components/schemas/AppointmentBase'
      - required:
        - patient_id
        - doctor_id
        - end_time
        - start_time
        - approved_by_patient
        - visited
        type: object
        properties:
          approved_by_patient:
            type: boolean
            description: whether the assigned patient approved that he/she will arrive
            example: true
            default: false
          visited:
            type: boolean
            description: whether the assigned patient come to the appointment
            default: false
    IdHolder:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
    PatientIdHolder:
      required:
      - patient_id
      type: object
      properties:
        patient_id:
          type: integer
          format: int32
          example: 25